+++ src/constants/index.ts
@@ -1,6 +1,6 @@
-export * from './numbers.const';
 export * from './numbers';
 export * from './generals.const';
+export * from './toolbar.const';
 export * from './typos.const';
 export * from './defaults.const';
 export * from './aiModels.const';